Fixed Bug with UBC Course Section Detail Scraper
What was the bug?
When scrapping data - if course had two sections listed in the same section it would concatenate the name together twice.

Eg.

CPSC 210 104 (Lecture)
Term	Day	Start Time	End Time	Building	Room
1	Wed	16:00	18:00	West Mall Swing Space	121
1	Fri	16:00	17:00	West Mall Swing Space	121

This would turn into:
"West Mall Swing SpaceWest Mall Swing Space"
"Room: 121121"

This is sloppy and needs to be Fixed

How was the bug fixed?
When taking the data from the scrapped website I choose to just take the first room that appears because it is the same room for both sections (in the scenarios I have seen).